backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-dotenv==1.0.1
PyPDF2==3.0.1
numpy==1.26.4
tiktoken==0.5.2
tqdm==4.66.2
pandas==2.2.2
openai==1.76.0
azure-storage-blob==12.19.0
azure-core==1.30.1
requests==2.31.0
httpx==0.28.1
typing-extensions>=4.11.0,<5.0.0